need signed comparison now to match buflines
svn:revision:54953
svn:branch:trunk
svn:account:brlcad

+++ src/util/bwcrop.c
@@ -246,7 +246,7 @@ main(int argc, char **argv)
     init_buffer(scanlen);
 
     /* Check for silly buffer syndrome */
-    if ((unsigned)abs((int)(ury - uly)) > buflines/2 || (unsigned)abs((int)(lry - lly)) > buflines/2) {
+    if ((ssize_t)abs((int)(ury - uly)) > buflines/2 || (ssize_t)abs((int)(lry - lly)) > buflines/2) {
 	fprintf(stderr, "bwcrop: Warning: You are skewing enough in the y direction\n");
 	fprintf(stderr, "bwcrop: relative to my buffer size that I will exhibit silly\n");
 	fprintf(stderr, "bwcrop: buffer syndrome (two replacements per scanline).\n");
